chore: Move process player data to parsing module
